Moved deserialization to EventReader component.
