+++ EventReader/src/Read/IDataSource.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace EventReader
+namespace EventReader.Read
 {
     public interface IDataSource
     {
