+++ EventReader/src/Read/SqlSource.cs
@@ -2,7 +2,7 @@
 using System.Data;
 using System.Data.SqlClient;
 
-namespace EventReader
+namespace EventReader.Read
 {
     public class SqlSource : IDataSource
     {
